Change folder hierarchy according to the guidelines
