+++ extensions/CUDA/src/BeamAdapter/CUDA/config.h.in
