+++ extensions/CUDA/src/BeamAdapter/CUDA/init.cpp
@@ -19,11 +19,49 @@
 *                                                                             *
 * Contact information: contact@sofa-framework.org                             *
 ******************************************************************************/
-#pragma once
-
-#include <BeamAdapter.CUDA/config.h>
-
+#include <BeamAdapter/CUDA/init.h>
+#include <BeamAdapter/initBeamAdapter.h>
+#include <sofa/core/ObjectFactory.h>
 namespace beamadapter::cuda
 {
-SOFA_BEAMADAPTER_CUDA_API void init();
+
+extern "C" {
+    SOFA_EXPORT_DYNAMIC_LIBRARY void initExternalModule();
+    SOFA_EXPORT_DYNAMIC_LIBRARY const char* getModuleName();
+    SOFA_EXPORT_DYNAMIC_LIBRARY const char* getModuleVersion();
+    SOFA_EXPORT_DYNAMIC_LIBRARY const char* getModuleComponentList();
+}
+
+void initExternalModule()
+{
+    init();
+}
+
+const char* getModuleName()
+{
+    return MODULE_NAME;
+}
+
+const char* getModuleVersion()
+{
+    return MODULE_VERSION;
+}
+
+void init()
+{
+    static bool first = true;
+    if (first)
+    {
+        sofa::component::initBeamAdapter();
+        first = false;
+    }
+}
+
+const char* getModuleComponentList()
+{
+    /// string containing the names of the classes provided by the plugin
+    static std::string classes = sofa::core::ObjectFactory::getInstance()->listClassesFromTarget(MODULE_NAME);
+    return classes.c_str();
+}
+
 } // namespace beamadapter::cuda
