+++ extensions/CUDA/src/BeamAdapter/CUDA/init.h
@@ -19,49 +19,11 @@
 *                                                                             *
 * Contact information: contact@sofa-framework.org                             *
 ******************************************************************************/
-#include <BeamAdapter.CUDA/init.h>
-#include <BeamAdapter/initBeamAdapter.h>
-#include <sofa/core/ObjectFactory.h>
+#pragma once
+
+#include <BeamAdapter/CUDA/config.h>
+
 namespace beamadapter::cuda
 {
-
-extern "C" {
-    SOFA_EXPORT_DYNAMIC_LIBRARY void initExternalModule();
-    SOFA_EXPORT_DYNAMIC_LIBRARY const char* getModuleName();
-    SOFA_EXPORT_DYNAMIC_LIBRARY const char* getModuleVersion();
-    SOFA_EXPORT_DYNAMIC_LIBRARY const char* getModuleComponentList();
-}
-
-void initExternalModule()
-{
-    init();
-}
-
-const char* getModuleName()
-{
-    return MODULE_NAME;
-}
-
-const char* getModuleVersion()
-{
-    return MODULE_VERSION;
-}
-
-void init()
-{
-    static bool first = true;
-    if (first)
-    {
-        sofa::component::initBeamAdapter();
-        first = false;
-    }
-}
-
-const char* getModuleComponentList()
-{
-    /// string containing the names of the classes provided by the plugin
-    static std::string classes = sofa::core::ObjectFactory::getInstance()->listClassesFromTarget(MODULE_NAME);
-    return classes.c_str();
-}
-
+SOFA_BEAMADAPTER_CUDA_API void init();
 } // namespace beamadapter::cuda
